pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ominix-tts"
version = "0.1.0"
description = "Ominix TTS: A multilingual TTS system"
readme = "README.md"
authors = [{name = "Hongbing Li", email = "cshbli@hotmail.com"}]
license = {file = "LICENSE"}
requires-python = ">=3.9"
dependencies = [
    "cn2an",
    "fast_langdetect>=0.3.0",
    "ffmpeg-python",
    "g2p_en",
    "gradio>=4.0,<=4.24.0",
    "huggingface_hub>=0.13",
    "jieba",
    "jieba_fast",
    "librosa>=0.9.2",
    "matplotlib",
    "numpy>=1.23.4,<2.0.0",
    "peft",
    "pypinyin",
    "pytorch-lightning>2.0",
    "split-lang",
    "torchaudio",
    "tqdm",
    "transformers>=4.43",
    "wordsegment",
    "x_transformers",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/cshbli/Ominix-TTS"

# Include data files in the package
[tool.setuptools.package-data]
"ominix_tts.tools.i18n" = ["locale/*.json"]
"ominix_tts.text" = [
    "opencpop-strict.txt", 
    "cmudict.rep", 
    "cmudict-fast.rep", 
    "engdict-hot.rep",
    "engdict_cache.pickle",
    "namedict_cache.pickle"
]
